pyproject.toml
[tool.poetry]
name = "microservices-films-remout-auth"
version = "0.1.0"
description = ""
authors = ["vladimirmac <vladimirbodrin89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
pydantic = {version = "2.6", extras = ["email"]}
asyncpg = "^0.29.0"
grpcio-tools = "^1.62.2"
fastapi-debug-toolbar = "^0.6.2"
tortoise-orm = "^0.20.0"
python-dotenv = "^1.0.1"
aerich = "^0.7.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"


[tool.aerich]
tortoise_orm = "src.config.settings.DATABASE_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
